[refactor] Relocate 'serve_application' and 'rewrite_hyperlinks' helper functions to the tests.utils module.
